Added gun traps
Added gun traps.  They shoot endlessly if they see something, and the
bullets travel and hopefully hit something.  The bullets move fairly
slowly.

Improvements could be made, as follows:
1) Bullets should use true collision detection (did I hit that guy
during the last second?) instead of intersection detection (am I inside
that guy right at this instant?).  Right now it doesn't matter, because
the bullets are moving so slowly that it would be impossible to go
completely through an enemy, but fixing this would let them move faster
and cope with smaller enemies.

2) Remove X-Ray vision.  That is, the gun fires if there's anything in
front of it, but it doesn't check whether there's a wall between the gun
and the enemy.  The bullet wouldn't pierce the wall, but it shouldn't be
fired at all.  Note that the current grid doesn't show this problem.

3) Maybe aim a little bit?  Firing in straight lines is eh