Enemies now drop cash and inflict damage
There is now a limited amount of health (3) and one is lost whenever an
enemy hits a goal.  When an enemy is killed, it drops some cash as well.

Buying and damage has be retuned so that it is possible to have a
sustainable trap setup without losing any lives, but you have to be
quick :)