Added winning, fixed wave tracker
Just what it says!  You're the best winner ever!

Also, moved a lot of the functionality of GameUpdater (which had too
much to do with the ArenaMap) into an ArenaManager class.  The Map
should now be a lot more constant, with the management done above.

The references are now hierarchical:

TDGame owns GameUpdater
GameUpdater owns ArenaManager, PlayerHUD, ButtonPanel
ArenaManager owns ArenaMap, Trap[], Enemy[], WaveTracker

The only violation is that GameUpdater draws WaveTracker separately from
the rest of the ArenaManager stuff, since it's not a contigous block,
and GameUpdater controls the visuals of the game screen.