Can now build walls
Walls now work.  Two things to be dealt with:

1) Enemies should prevent the construction of new walls; right now they
just kind of glitch out, disappear, and damage the player (they are
considered to be "at a goal" when they can't figure out where to go).

2) Something interesting should happen when pathing is broken- probably
enemies digging randomly?  In any case enemies currently glitch out and
hurt the player, like in case (1)