Wall Issue (1) resolved
You can no longer place walls over enemies, or where they are partially
in the way.

The solution is a little rigid; all enemies now have a CurrentSquare and
NextSquare, which they basically completely block.  It's not a bad
solution but it might be strange for the player if an enemy seems to
block more area than it should.  It also doesn't play nice with any sort
of blowback towers I might someday implement, so this system might not
last.