Knockback removed :(
Knockback is removed, and the heavy cannon is back to shitty
non-balance.

The mechanic isn't a good fit for the game, because enemies move very
rigidly from square to square, but smoothly between them.  Knockback
would hit something in the middle of that transition and screw up the
rigid motion.

Allowing for this could be done, but we'd have to change the engine
around a fair amount, and the result would be a game that feels a lot
less like ... well like this.