Added pits!
Pits are pretty cool; they block pathing, but bullets can shoot over
them.  It's a good way to expand the area of influence of a poison
tower, etc.

Right now they're strictly better than walls and much cheaper.  Later
there will be flying enemies and that'll be something to deal with :)