Can now get previews of objects in play
The interface is a little awkward: if you're in "buy mode" which needs
to be a little more obvious, your clicks need to be used for buying
traps, instead of clicking enemies.  The only way to end "buy mode" is
to hit clear, which doesn't even really make sense.

But the functionality is there, it's worth a commit.