Achievements implemented
Which is actually a pain in the ass, and now there's little reporting
methods hovering around everywhere in weird places.  Funny achievements
for funny things are less elegant than just "beat the game"
achievements.

Last thing left to do is give popups in the game when you get an
achievement.